Fixes to gpytorch to be compatible with PyTorch master
This retains compatiblity with PyTorch 0.3.

The file gpytorch/utils/function_factory.py deserves some attention: The
`t_matmul_closure_factory` does not always return the correct shape for
the gradient. This is a bug: gradients should have the exact same shape
as the differnetiated variable. PyTorch is missing some checks for this
behavior, but future versions of PyTorch are likely to raise an
exception if shapes don't exactly match.