Register torch.isclose
This is called in some places inside the torch.distributions code when `validate_args=True`.
This is a pretty basic implementation and we probably don't want to do this kind of things if the linear operators are very large since `islose` returns a tensor of the same size as the (often implicitly represented) operator.