Introduce reshape
This function is used ubiquitously with tensors, so in order to make `LinearOperator` a drop-in replacement for tensors we need to use it. `reshape` (other than view) will reallocate new memory if neessary - this concept doesn't really make a to of sense for `LinearOperator`s so a straightforward thing to do is just to make this an alias for `expand` and not allow reshaping the last two (non-batch) dimensions for now.